% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-refequal.R
\name{:=}
\alias{:=}
\title{Pipe operator}
\usage{
lhs := rhs
}
\arguments{
\item{lhs}{A value.}

\item{rhs}{A function call.}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
See \code{data.table::\link[data.table::=]{:=}} for details.
}
\keyword{internal}
