% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{download}
\alias{download}
\title{Функция обновляет базу данных для определенных тикеров или источников.
Если тикеры и источники не указаны, производится обновление всей базы данных, запись сырых данных в папку data/raw (файлы ticker.csv с тремя колонками: date, value, update_date),
запись использованных excel-файлов из Росстата в папку data/raw_excel (файлы .xls, .xlsx в исходном виде). Обратите внимание: в папку raw происходит запись таких комбинаций (date, value),
которые до этого не встречались в таблице. Некоторые переменные, в частности индексы OECD и Индекс глобальной экономической активности, каждый месяц пересчитываются для всех значений.
Это означает, что соответствующие таблицы в папке raw будут полностью обновляться, однако старые значения удаляться не будут. Также следует учесть, что папка data/raw_excel
потенциально может занимать достаточно много места,
так как в ней хранятся все исходные excel-файлы, когда-либо использованные для скачивания данных с Росстата.
Папка raw состоит из нескольких подпапок, названных так же, как называются отдельные таблицы, которые можно получить с Росстата \code{macroparsing::rosstat_tables}.
В каждой из подпапок сохраняются версии исходных файлов, названные по датам обновления файла на сайте Росстата.
В папке transform для каждого тикера представлена трансформированная версия временного ряда (таблица date, value): во-первых, для каждой даты есть только одно значение, соответствующее
самому актуальном значению ряда. Во-вторых, дневные временные ряды переведены в среднее с начала месяца значение. В-третьих, те ряды, которые исходно представляются
отностельно прошлого месяца (в частности, ИПЦ), переведены в цепные индексы.}
\usage{
download(tickers = NULL, sources = NULL, raw = TRUE, transform = TRUE)
}
\arguments{
\item{tickers}{строковый вектор тикеров из таблицы \code{macroparsing::show.variables()}}

\item{sources}{строковый вектор источников из таблицы \code{macroparsing::sources}}

\item{raw}{логическое выражение, если верно, то происходит запрос данных из внешних источников и обновляется содержимое папок data/raw/ и data/raw_excel.}

\item{transform}{логическое выражение, если верно, то на основе содержмого папки data/raw/ данные трансформируются и заполняется содержимое папки data/transform}
}
\description{
Функция обновляет базу данных для определенных тикеров или источников.
Если тикеры и источники не указаны, производится обновление всей базы данных, запись сырых данных в папку data/raw (файлы ticker.csv с тремя колонками: date, value, update_date),
запись использованных excel-файлов из Росстата в папку data/raw_excel (файлы .xls, .xlsx в исходном виде). Обратите внимание: в папку raw происходит запись таких комбинаций (date, value),
которые до этого не встречались в таблице. Некоторые переменные, в частности индексы OECD и Индекс глобальной экономической активности, каждый месяц пересчитываются для всех значений.
Это означает, что соответствующие таблицы в папке raw будут полностью обновляться, однако старые значения удаляться не будут. Также следует учесть, что папка data/raw_excel
потенциально может занимать достаточно много места,
так как в ней хранятся все исходные excel-файлы, когда-либо использованные для скачивания данных с Росстата.
Папка raw состоит из нескольких подпапок, названных так же, как называются отдельные таблицы, которые можно получить с Росстата \code{macroparsing::rosstat_tables}.
В каждой из подпапок сохраняются версии исходных файлов, названные по датам обновления файла на сайте Росстата.
В папке transform для каждого тикера представлена трансформированная версия временного ряда (таблица date, value): во-первых, для каждой даты есть только одно значение, соответствующее
самому актуальном значению ряда. Во-вторых, дневные временные ряды переведены в среднее с начала месяца значение. В-третьих, те ряды, которые исходно представляются
отностельно прошлого месяца (в частности, ИПЦ), переведены в цепные индексы.
}
\examples{
# РћР±РЅРѕРІРёС‚СЊ РІСЃРµ РґРѕСЃС‚СѓРїРЅС‹Рµ РІСЂРµРјРµРЅРЅС‹Рµ СЂСЏРґС‹
download()
# РћР±РЅРѕРІРёС‚СЊ РРџР¦ (С‚РёРєРµСЂ cpi)
download(tickers = c("cpi"))
# РћР±РЅРѕРІРёС‚СЊ РІСЃРµ СЂСЏРґС‹ РёР· OECD (С‚РёРєРµСЂ РёСЃС‚РѕС‡РЅРёРєР° oecd)
download(sources = c("oecd"))
# РћР±РЅРѕРІРёС‚СЊ С‚РѕР»СЊРєРѕ СЃС‹СЂС‹Рµ Р·РЅР°С‡РµРЅРёСЏ Рё РЅРµ С‚СЂР°РЅСЃС„РѕСЂРјРёСЂРѕРІР°С‚СЊ РёС…
download(transform = FALSE)
#' РћР±РЅРѕРІРёС‚СЊ С‚РѕР»СЊРєРѕ РїР°РїРєСѓ transform, РЅРµ Р·Р°РіСЂСѓР¶Р°СЏ СЂСЏРґС‹ РёР· РІРЅРµС€РЅРёС… РёСЃС‚РѕС‡РЅРёРєРѕРІ
download(raw = FALSE)
}
