% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{fill.folder}
\alias{fill.folder}
\title{Update specified folder}
\usage{
fill.folder(tickers = NULL, sources = NULL, folder = c("raw", "tf"))
}
\arguments{
\item{tickers}{character, список допустимых тикеров см. \code{rmedb::show.variables()}}

\item{sources}{character, список допустимых имен источников см. \code{rmedb::sources}}

\item{type}{character: "raw", "transform"}
}
\value{

}
\description{
Обновляет файлы для указанных тикеров и источников в указанной папке type.
}
\keyword{internal}
