% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{get.variables.df}
\alias{get.variables.df}
\title{Variables dataframe}
\usage{
get.variables.df(tickers = NULL, sources = NULL)
}
\arguments{
\item{tickers}{character, список допустимых тикеров, см. \link[rmedb]{show.variables}}

\item{sources}{character, список допустимых имен источников, см. \link[rmedb]{sources}}
}
\value{
data.frame, tibble
}
\description{
Возвращает data.frame, отфильтрованный по указанным тикерам и источникам.
}
\keyword{(не}
\keyword{internal}
\keyword{internal.}
\keyword{Если}
\keyword{а}
\keyword{были}
\keyword{временные}
\keyword{все}
\keyword{других}
\keyword{зависимые}
\keyword{запись}
\keyword{значений}
\keyword{и}
\keyword{из}
\keyword{источника}
\keyword{источнику}
\keyword{их}
\keyword{к}
\keyword{на}
\keyword{напрямую,}
\keyword{новых}
\keyword{обновления}
\keyword{основе}
\keyword{после}
\keyword{принадлежащие}
\keyword{происходит}
\keyword{рассчитываемые}
\keyword{рядов),}
\keyword{ряды,}
\keyword{скачиваемые}
\keyword{скачиваются}
\keyword{сначала}
\keyword{тикеров}
\keyword{тикеры,}
\keyword{то}
\keyword{только}
\keyword{указаны}
