% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_parsing.R
\name{initialize,cbr-method}
\alias{initialize,cbr-method}
\title{Title}
\usage{
\S4method{initialize}{cbr}(
  .Object,
  ticker,
  observation_start,
  previous_date_till,
  date_from,
  ts,
  url,
  freq,
  cbr_ticker
)
}
\arguments{
\item{cbr}{}
}
\value{

}
\description{
Title
}
