% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{class}
\name{parsed_ts-class}
\alias{parsed_ts-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{ticker}}{character.}

\item{\code{observation_start}}{Date.}

\item{\code{previous_date_till}}{Date.}

\item{\code{date_from}}{Date.}

\item{\code{ts}}{data.frame.}
}}

