% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.environment}
\alias{set.environment}
\title{Title}
\usage{
set.environment(
  path = "C:/Users/Mikhail Gareev/Documents/macroparsing_usage",
  fredr_api_key = "aaa"
)
}
\arguments{
\item{path}{character.}

\item{fredr_api_key}{character.}
}
\value{

}
\description{
Title
}
