% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visual.R
\name{show.variables}
\alias{show.variables}
\title{Show variables}
\usage{
show.variables(additional = FALSE, russificate = FALSE, url_as_href = FALSE)
}
\arguments{
\item{additional}{logical показывать дополнительные колонк}

\item{russificate}{logical русифицировать названия источников и периодов}

\item{url_as_href}{logical сделать url-ссылки кликабельными}
}
\value{
data.frame
}
\description{
Возвращает data.frame со справочной информацией по переменным, которые доступны для скачивания в пакете.
}
