% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{sources}
\alias{sources}
\title{Sources}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6 rows and 4 columns.
}
\usage{
sources
}
\description{
Таблица с источниками данных.
}
\details{
\itemize{
\item source. Тикер источника
\item url. Адрес
\item name_eng. Название на английском
\item name_rus. Названием на русском
}
}
\keyword{datasets}
