% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{variables}
\alias{variables}
\title{Here is the title of the dataset}
\format{
data.table
}
\source{
This is a link to
\href{http://www.xxx.xxx.com/xxx.csv}{the origin data}
}
\usage{
data("var_list")
}
\description{
More description of the dataset is here
}
\keyword{datasets}
