% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{download}
\alias{download}
\title{Title}
\usage{
download(
  tickers = NULL,
  sources = NULL,
  use_future = FALSE,
  raw = TRUE,
  transform = TRUE
)
}
\arguments{
\item{transform_and_deseason}{}
}
\value{

}
\description{
Title
}
