% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{download}
\alias{download}
\title{Update database}
\usage{
download(tickers = NULL, sources = NULL, raw = TRUE, transform = TRUE)
}
\arguments{
\item{tickers}{character список допустимых тикеров см. см. \link[rmedb]{show.variables}}

\item{sources}{character список допустимых имен источников см. \link[rmedb]{sources}}

\item{raw}{logical, при TRUE происходит запрос данных из внешних источников и обновляется содержимое папок data/raw/ и data/raw_excel}

\item{transform}{logical, при TRUE на основе содержмого папки data/raw/ данные трансформируются и заполняется содержимое папки data/tf}
}
\description{
Получает обновления данных для указанных временных рядов (ticker) или всех временных рядов из указанных источников (source) и записывает их в рабочую директорию пакета.
Если тикеры и источники не указаны, производится обновление всей базы данных.
}
