% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.environment}
\alias{set.environment}
\title{Create database directory}
\usage{
set.environment(path, fredr_api_key)
}
\arguments{
\item{path}{character путь к рабочей директории}

\item{fredr_api_key}{character ключ FREDR API}
}
\description{
Создает по указанному адресу папку и необходимые подпапки для дальнейшей записи туда базы данных,
а также создает в домашней директории файл .Renviron с глобальными переменными окружения.
}
