% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-config.R
\name{set.environment}
\alias{set.environment}
\title{Set environment variables}
\usage{
set.environment(path, fredr_api_key)
}
\arguments{
\item{path}{character путь к рабочей директории}

\item{fredr_api_key}{character ключ FREDR API}
}
\description{
Создает в домашней директории файл .Renviron с глобальными переменными окружения.
}
